Another huge overhaul of the protocol design.
The gather_information paradigm wasn't working, so it was scrapped.
Commands now all begin with an authentication step, which is
currently blanketly acknowledged, since the key is in shared
memory. This will be switched when necessary. Commands now
transmit/receieve data only when necessary for their function.
Functions related to buffet_t are now more aptly named.